Move return button click handler  (#147)
