+++ frontend/app/components/me/LoanSelectedDialog.tsx
@@ -1,15 +1,21 @@
 import { Button, Dialog, Stack } from '@mantine/core';
+import { useSubmit } from '@remix-run/react';
 import { useAtom } from 'jotai';
 import { selectedLoanAtom } from '~/stores/loanAtom';
 
-interface LoanSelectedDialogProps {
-	handleReturnButtonClick: () => void;
-}
-
-const LoanSelectedDialog = ({
-	handleReturnButtonClick,
-}: LoanSelectedDialogProps) => {
+const LoanSelectedDialog = () => {
 	const [selectedLoan, setSelectedLoan] = useAtom(selectedLoanAtom);
+
+	const submit = useSubmit();
+
+	const handleReturnButtonClick = () => {
+		submit(JSON.stringify({ selectedLoan: selectedLoan }), {
+			action: '/home/me?index',
+			method: 'PATCH',
+			encType: 'application/json',
+		});
+		setSelectedLoan([]);
+	};
 
 	return (
 		<Dialog
